requirements.txt
nornir==3.3.0
nornir-napalm==0.4.0
nornir-scrapli==2023.7.30
pyats==23.7
pynetbox==7.2.0
dnspython==2.6.1
genie==23.7
GitPython==3.1.32
motor==3.4.0
netaddr==1.3.0

